feat: finalize big-party-IV and add icons
