+++ components/icons.tsx
@@ -1,0 +1,67 @@
+type Props = JSX.IntrinsicElements['svg']
+
+export const LinkedInIcon = (props: Props) => {
+  return (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>LinkedIn icon</title>
+      <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z" />
+    </svg>
+  )
+}
+
+export const GithubIcon = (props: Props) => {
+  return (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>GitHub icon</title>
+      <path d="M12 .297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C4.422 18.07 3.633 17.7 3.633 17.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 22.092 24 17.592 24 12.297c0-6.627-5.373-12-12-12" />
+    </svg>
+  )
+}
+
+export const TwitterIcon = (props: Props) => {
+  return (
+    <svg
+      role="img"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <title>Twitter icon</title>
+      <path d="M23.954 4.569c-.885.389-1.83.654-2.825.775 1.014-.611 1.794-1.574 2.163-2.723-.951.555-2.005.959-3.127 1.184-.896-.959-2.173-1.559-3.591-1.559-2.717 0-4.92 2.203-4.92 4.917 0 .39.045.765.127 1.124C7.691 8.094 4.066 6.13 1.64 3.161c-.427.722-.666 1.561-.666 2.475 0 1.71.87 3.213 2.188 4.096-.807-.026-1.566-.248-2.228-.616v.061c0 2.385 1.693 4.374 3.946 4.827-.413.111-.849.171-1.296.171-.314 0-.615-.03-.916-.086.631 1.953 2.445 3.377 4.604 3.417-1.68 1.319-3.809 2.105-6.102 2.105-.39 0-.779-.023-1.17-.067 2.189 1.394 4.768 2.209 7.557 2.209 9.054 0 13.999-7.496 13.999-13.986 0-.209 0-.42-.015-.63.961-.689 1.8-1.56 2.46-2.548l-.047-.02z" />
+    </svg>
+  )
+}
+
+export const NgPartyIcon = (props: Props) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill-rule="evenodd"
+      stroke-linejoin="round"
+      stroke-miterlimit="2"
+      clip-rule="evenodd"
+      viewBox="0 0 1195 1191"
+      {...props}
+    >
+      <title>ngParty logo</title>
+      <defs />
+      <path fill="none" d="M-236-238h1666.67v1666.67H-236z" />
+      <path fill="#212158" d="M-1.646-3.648h1197.96v1197.96H-1.646z" />
+      <path
+        fill="#fff"
+        fill-rule="nonzero"
+        d="M431.86 513.144V689h50.673V567.878c11.653-8.298 24.895-12.889 35.136-12.889 14.301 0 20.481 7.592 20.481 21.364V689h50.673V560.639c0-33.194-18.362-51.026-50.32-51.026-22.247 0-43.611 10.24-60.561 22.247l-2.648-18.716H431.86zM737.489 470.769c-2.649 12.889-8.299 31.958-13.772 42.728-8.298-2.472-17.656-3.884-27.367-3.884-45.2 0-74.509 23.482-74.509 57.559v14.125c0 20.834 10.946 37.608 29.662 47.495-11.3 7.769-21.541 19.599-21.541 33.9 0 11.477 7.24 20.835 16.95 26.308-13.771 7.416-28.073 20.481-28.073 39.02 0 28.78 24.366 45.024 65.858 45.024h30.015c41.316 0 74.863-17.657 74.863-55.088 0-38.49-28.603-58.089-64.622-58.089h-43.434c-7.239 0-12.007-3.178-12.007-9.004 0-4.591 3.885-9.711 8.652-13.772 5.65 1.059 11.83 1.589 18.186 1.589 48.378 0 76.628-25.602 76.628-61.091v-6.886c0-18.715-6.18-33.193-20.481-44.493 10.77-10.241 22.776-26.838 29.662-42.905l-44.67-12.536zm-66.564 95.873c0-12.889 8.651-24.718 26.484-24.718 19.246 0 26.485 14.478 26.485 28.779v6.71c0 14.301-7.239 28.25-26.485 28.25-19.245 0-26.484-13.772-26.484-28.25v-10.771zm49.261 130.303c13.772 0 24.895 5.121 24.895 20.482 0 17.832-17.126 19.245-34.783 19.245h-18.892c-17.48 0-30.192-4.061-30.192-18.539 0-10.064 7.769-17.48 14.831-21.364 3.178.176 6.356.176 9.711.176h34.43z"
+      />
+    </svg>
+  )
+}
