+++ components/index.ts
@@ -1,3 +1,4 @@
 export * from './amp-style-boilerplate'
 export * from './amp-style-keyframes'
 export * from './amp-style-custom'
+export * from './icons'
