+++ conferences/big-party-iv/components/about.tsx
@@ -1,0 +1,18 @@
+export const About = () => {
+  return (
+    <section className="teaser ngparty-conf-section">
+      <h2>ngBigParty IV</h2>
+      <p>
+        <span>
+          <strong>March 22</strong> in Prague.
+        </span>
+        <span>
+          <strong>6</strong> Talks.
+        </span>
+        <span>
+          <strong>1</strong> Panel.
+        </span>
+      </p>
+    </section>
+  )
+}
