+++ conferences/big-party-iv/components/faq.tsx
@@ -1,0 +1,19 @@
+export const FAQ = () => {
+  return (
+    <section className="faq ngparty-conf-section">
+      <div className="wrapper">
+        <h3 id="faq">Frequently Asked Questions</h3>
+        <dl>
+          <dt>Does it cost anything?</dt>
+          <dd>No - tickets are free of charge.</dd>
+
+          <dt>Who is the event for?</dt>
+          <dd>Web developers who seek for new technologies and have a fun</dd>
+
+          <dt>When will I hear back if I have a spot?</dt>
+          <dd>We'll try to confirm you within less than a week!</dd>
+        </dl>
+      </div>
+    </section>
+  )
+}
