+++ conferences/big-party-iv/index.ts
@@ -1,3 +1,3 @@
 export * from './data'
-export * from './styles'
 export * from './layout'
+export * from './components'
